Restructurered rest of the sources
