+++ src/beetl/sources/faker/__init__.py
@@ -1,0 +1,2 @@
+from .faker_config import FakerConfig, FakerConfigArguments
+from .faker_source import FakerSource
