+++ src/beetl/sources/itop/__init__.py
@@ -1,0 +1,4 @@
+# re-export
+from .itop_config import ItopConfig, ItopConfigArguments
+from .itop_source import ItopSource
+from .itop_sync import ItopSync, ItopSyncArguments
