+++ src/beetl/sources/postgresql/__init__.py
@@ -1,0 +1,4 @@
+# re-export
+from .postgresql_config import PostgresConfig, PostgresConfigArguments
+from .postgresql_source import PostgresSource
+from .postgresql_sync import PostgresSync, PostgresSyncArguments
