+++ src/beetl/sources/static/__init__.py
@@ -1,0 +1,2 @@
+from .static_config import StaticConfig, StaticConfigArguments
+from .static_source import StaticSource
